# apps/api/docs/openapi_v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CollegeERPAPI",
    "version": "1.0.0",
    "description": "\nCollege ERP System API\n\n**Generated:** 2026-02-03T12:39:45.480081\n\n**Locked Domains:**\n- Admission (v1.0.0) - Contracts locked with strict validation\n\n**Status:**\n- \u2705 Admission: Ready for frontend integration\n- \u23f3 Finance: In progress\n- \u23f3 Academic: In progress\n- \u23f3 Student: In progress\n\n**Breaking Changes Policy:**\n- Enum additions: Safe (backward compatible)\n- New optional fields: Safe\n- Required field changes: Requires migration + deprecation notice\n- Enum removals: 6-month deprecation period\n- Field type changes: Major version bump\n        "
  },
  "paths": {
...
    "/api/v1/campus/library/library/books": {
      "post": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "Create Book",
        "description": "Add a new book to the library catalog",
        "operationId": "create_book_api_v1_campus_library_library_books_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "List Books",
        "description": "List books in the catalog with filters",
        "operationId": "list_books_api_v1_campus_library_library_books_get",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Author"
            }
          },
          {
            "name": "available_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Available Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  },
                  "title": "Response List Books Api V1 Campus Library Library Books Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/library/library/members": {
      "post": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "Create Membership",
        "description": "Create a new library membership",
        "operationId": "create_membership_api_v1_campus_library_library_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryMember"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campus/library/library/members/{member_id}/statistics": {
      "get": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "Get Member Statistics",
        "description": "Get circulation statistics for a member",
        "operationId": "get_member_statistics_api_v1_campus_library_library_members__member_id__statistics_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/library/library/issue": {
      "post": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "Issue Book",
        "description": "Issue a book to a member",
        "operationId": "issue_book_api_v1_campus_library_library_issue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookIssueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookIssue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/campus/library/library/return/{issue_id}": {
      "post": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "Return Book",
        "description": "Return an issued book",
        "operationId": "return_book_api_v1_campus_library_library_return__issue_id__post",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookIssue"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campus/library/library/issues": {
      "get": {
        "tags": [
          "campus",
          "Campus - Library",
          "Library Circulation"
        ],
        "summary": "List Issues",
        "description": "List book issues with filters",
        "operationId": "list_issues_api_v1_campus_library_library_issues_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IssueStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "overdue_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Overdue Only"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookIssue"
                  },
                  "title": "Response List Issues Api V1 Campus Library Library Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Book": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "isbn": {
            "type": "string",
            "title": "Isbn"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "total_copies": {
            "type": "integer",
            "title": "Total Copies",
            "default": 1
          },
          "available_copies": {
            "type": "integer",
            "title": "Available Copies",
            "default": 1
          },
          "status": {
            "$ref": "#/components/schemas/BookStatus",
            "default": "AVAILABLE"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "isbn"
        ],
        "title": "Book",
        "description": "Intellectual Resource - Library Collection Item"
      },
      "BookCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "isbn": {
            "type": "string",
            "title": "Isbn"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "total_copies": {
            "type": "integer",
            "title": "Total Copies",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "isbn"
        ],
        "title": "BookCreate"
      },
      "BookIssue": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "library_member_id": {
            "type": "integer",
            "title": "Library Member Id"
          },
          "issue_date": {
            "type": "string",
            "format": "date",
            "title": "Issue Date"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "return_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Date"
          },
          "status": {
            "$ref": "#/components/schemas/IssueStatus",
            "default": "ISSUED"
          },
          "is_returned": {
            "type": "boolean",
            "title": "Is Returned",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "library_member_id",
          "due_date"
        ],
        "title": "BookIssue",
        "description": "Loan Record - Tracking intellectual resource usage"
      },
      "BookIssueRequest": {
        "properties": {
          "book_id": {
            "type": "integer",
            "title": "Book Id"
          },
          "member_id": {
            "type": "integer",
            "title": "Member Id"
          },
          "due_days": {
            "type": "integer",
            "title": "Due Days",
            "default": 14
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "member_id"
        ],
        "title": "BookIssueRequest"
      },
      "BookStatus": {
        "type": "string",
        "enum": [
          "AVAILABLE",
          "ISSUED",
          "LOST",
          "DAMAGED"
        ],
        "title": "BookStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IssueStatus": {
        "type": "string",
        "enum": [
          "ISSUED",
          "RETURNED",
          "OVERDUE",
          "LOST"
        ],
        "title": "IssueStatus"
      },
      "LibraryMember": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "member_type": {
            "$ref": "#/components/schemas/MemberType",
            "default": "STUDENT"
          },
          "card_number": {
            "type": "string",
            "title": "Card Number"
          },
          "max_books": {
            "type": "integer",
            "title": "Max Books",
            "default": 3
          },
          "books_issued_count": {
            "type": "integer",
            "title": "Books Issued Count",
            "default": 0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "ACTIVE"
          },
          "joined_date": {
            "type": "string",
            "format": "date",
            "title": "Joined Date"
          }
        },
        "type": "object",
        "required": [
          "card_number"
        ],
        "title": "LibraryMember",
        "description": "Library Patron - Access and Loan Limits"
      },
...
      "MemberType": {
        "type": "string",
        "enum": [
          "STUDENT",
          "FACULTY",
          "STAFF"
        ],
        "title": "MemberType"
      },
      "MembershipCreate": {
        "properties": {
          "student_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "member_type": {
            "$ref": "#/components/schemas/MemberType",
            "default": "STUDENT"
          },
          "max_books": {
            "type": "integer",
            "title": "Max Books",
            "default": 3
          }
        },
        "type": "object",
        "title": "MembershipCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}